<refactor> 패키지 변경 및 검증로직 분리
