--- src/main/java/kitchenpos/eatinorders/domain/OrderType.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorders.domain;
-
-public enum OrderType {
-    DELIVERY, TAKEOUT, EAT_IN
-}
