--- src/main/java/kitchenpos/eatinorders/tobe/domain/model/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain.model;
-
-public enum OrderStatus {
-    WAITING, ACCEPTED, SERVED, COMPLETED
-}
