+++ src/main/java/kitchenpos/tableorders/domain/JpaOrderRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.tableorders.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
+}
