+++ src/main/java/kitchenpos/tableorders/domain/JpaOrderTableRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.tableorders.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
+}
