+++ src/main/java/kitchenpos/tableorders/domain/OrderLineItem.java
@@ -1,0 +1,76 @@
+package kitchenpos.tableorders.domain;
+
+import kitchenpos.menus.domain.Menu;
+
+import javax.persistence.*;
+import java.math.BigDecimal;
+import java.util.UUID;
+
+@Table(name = "order_line_item")
+@Entity
+public class OrderLineItem {
+    @Column(name = "seq")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    private Long seq;
+
+    @ManyToOne(optional = false)
+    @JoinColumn(
+        name = "menu_id",
+        columnDefinition = "varbinary(16)",
+        foreignKey = @ForeignKey(name = "fk_order_line_item_to_menu")
+    )
+    private Menu menu;
+
+    @Column(name = "quantity", nullable = false)
+    private long quantity;
+
+    @Transient
+    private UUID menuId;
+
+    @Transient
+    private BigDecimal price;
+
+    public OrderLineItem() {
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public void setSeq(final Long seq) {
+        this.seq = seq;
+    }
+
+    public Menu getMenu() {
+        return menu;
+    }
+
+    public void setMenu(final Menu menu) {
+        this.menu = menu;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(final long quantity) {
+        this.quantity = quantity;
+    }
+
+    public UUID getMenuId() {
+        return menuId;
+    }
+
+    public void setMenuId(final UUID menuId) {
+        this.menuId = menuId;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(final BigDecimal price) {
+        this.price = price;
+    }
+}
