+++ src/main/java/kitchenpos/tableorders/domain/OrderRepository.java
@@ -1,0 +1,16 @@
+package kitchenpos.tableorders.domain;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderRepository {
+    Order save(Order order);
+
+    Optional<Order> findById(UUID id);
+
+    List<Order> findAll();
+
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
+}
+
