+++ src/main/java/kitchenpos/tableorders/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.tableorders.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
