+++ src/main/java/kitchenpos/tableorders/domain/OrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.tableorders.domain;
+
+public enum OrderType {
+    DELIVERY, TAKEOUT, EAT_IN
+}
