+++ src/main/java/kitchenpos/tableorders/tobe/domain/model/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.tableorders.tobe.domain.model;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
