+++ src/main/java/kitchenpos/tableorders/ui/OrderTableRestController.java
@@ -1,0 +1,50 @@
+package kitchenpos.tableorders.ui;
+
+import kitchenpos.tableorders.application.OrderTableService;
+import kitchenpos.tableorders.domain.OrderTable;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+
+@RequestMapping("/api/order-tables")
+@RestController
+public class OrderTableRestController {
+    private final OrderTableService orderTableService;
+
+    public OrderTableRestController(final OrderTableService orderTableService) {
+        this.orderTableService = orderTableService;
+    }
+
+    @PostMapping
+    public ResponseEntity<OrderTable> create(@RequestBody final OrderTable request) {
+        final OrderTable response = orderTableService.create(request);
+        return ResponseEntity.created(URI.create("/api/order-tables/" + response.getId()))
+            .body(response);
+    }
+
+    @PutMapping("/{orderTableId}/sit")
+    public ResponseEntity<OrderTable> sit(@PathVariable final UUID orderTableId) {
+        return ResponseEntity.ok(orderTableService.sit(orderTableId));
+    }
+
+    @PutMapping("/{orderTableId}/clear")
+    public ResponseEntity<OrderTable> clear(@PathVariable final UUID orderTableId) {
+        return ResponseEntity.ok(orderTableService.clear(orderTableId));
+    }
+
+    @PutMapping("/{orderTableId}/number-of-guests")
+    public ResponseEntity<OrderTable> changeNumberOfGuests(
+        @PathVariable final UUID orderTableId,
+        @RequestBody final OrderTable request
+    ) {
+        return ResponseEntity.ok(orderTableService.changeNumberOfGuests(orderTableId, request));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<OrderTable>> findAll() {
+        return ResponseEntity.ok(orderTableService.findAll());
+    }
+}
